tower mechanical
(as posted on the google group) the Santa Barbot Tower 2.0 is ready for
review!
its pretty similar in form to last year's barbot. As this is my first 3d
jigsaw design, i resisted temptations to get too crazy - but if this all
works fine maybe we can experiment with some more daring shapes in the
future.
See the sketchup model here, and the lasercutter design files here.
Also, uploaded to the Git repository.

Some notes:
-cost is about 25.46/ tower in an order of 6 (because of the geometry
the optimal order is a multiple of 6. unfortunately it goes up $3/tower
for an order of 8) . . I have a couple of ideas on ways this could be
reduced, but i'll save that for when there is more time for
experimentation. (things like lining up the positive and negative teeth
edges to pack the pieces more tightly and reduce the number of cuts)
-Leaving the opportunity open for Andrew's suggestion, the mounting
bracket on the base also has its own teeth. so if we want to make
additional connectors that allow the towers to snap into one another for
a fixed radius, the option is there
- I think some people were also interested in a miniature 'home' model?
a 1:2 scale version could be created with just a couple of clicks
- there are lots of colors available. are we going with RGB LEDs in the
base again (I'm assuming so) ? all the top baseplates should be clear,
and everything else can be another color. I put the baseplate files in a
separate eps file for this reason.
-self threading (.1006"diameter) 4-40 standoff holes on baseplate
-Each part is laser engraved with a number  corresponding to the label
in the sketchup model.
- we could do some other laser engraving too- maybe a logo on the
baseplate? our names on the back? i'm open to suggestions
- the sketchup model : shows assembled, exploded, and a combo view. if
you have any questions about assembly, let me know. it should be easy
though
-I feel fairly confident about the design but as mentioned, it is a
first attempt. There is a chance you will have to drill a few new holes
or cut something, but in the end it should work. Also I (intentionally)
did not account for the laser cut kerf. Meaning things should fit
together easily, but might need a bonding agent to stay put.